+++ src/core/index/sorted_set_doc_values_writer.rs
@@ -88,7 +88,7 @@ pub(crate) struct SortedSetDocValuesWriter {
 impl SortedSetDocValuesWriter {
   pub(crate) fn new(field_info: Arc<FieldInfo>, iw_bytes_used: SharedCounter) -> Result<Self> {
     let bytes_start_array =
-      DirectBytesStartArray::with_counter(DEFAULT_CAPACITY, iw_bytes_used.clone());
+      DirectBytesStartArray::with_counter(DEFAULT_CAPACITY as usize, iw_bytes_used.clone());
     let hash = BytesRefHash::from_bytes_start_array(DEFAULT_CAPACITY, bytes_start_array);
     let pending = PackedLongValues::delta_packed_long_values_builder_default(PackedInts::COMPACT)?;
     let docs_with_field = DocsWithFieldSet::new();
